requirements.txt
Django==5.2.1
psycopg2-binary==2.9.9
django-crispy-forms==2.4
crispy-bootstrap4==2024.10
djangorestframework==3.16.0
typing-extensions==4.13.2
dataclasses==0.6
pytz==2025.2
gunicorn==21.2.0
python-dotenv==1.0.1
drf-spectacular==0.28.0 